# 게시글 댓글 CRUD, 좋아요 기능 요청 처리 구현

+++ src/main/java/com/ssafy/wpwk/controller/BoardController.java
@@ -2,13 +2,19 @@ package com.ssafy.wpwk.controller;
 
 import com.ssafy.wpwk.model.Board;
 import com.ssafy.wpwk.service.BoardServiceImpl;
+import io.jsonwebtoken.Claims;
 import io.swagger.annotations.ApiOperation;
+import org.springframework.aop.AfterReturningAdvice;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
+
+import static com.ssafy.wpwk.utils.ExceptionUtil.isInValidAuthentication;
 
 @CrossOrigin(origins = {"*"}, maxAge = 6000)
 @RestController
@@ -19,13 +25,19 @@ public class BoardController {
 
     @ApiOperation(value = "게시글 생성")
     @PostMapping("/board")
-    public ResponseEntity<?> createBoard(@RequestBody Board board) {
+    public ResponseEntity<?> createBoard(@RequestBody Board board, Authentication authentication) {
+
+        if (isInValidAuthentication(authentication)) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         try {
             boardService.create(board);
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
+
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
@@ -57,8 +69,14 @@ public class BoardController {
 
     @ApiOperation(value = "ID를 이용한 게시글 상세 조회")
     @GetMapping("/board/{id}")
-    public ResponseEntity<?> findBoardById(@PathVariable("id") Long id) {
+    public ResponseEntity<?> findBoardById(@PathVariable("id") Long id, Authentication authentication) {
+
         Board board;
+
+        if (isInValidAuthentication(authentication)) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         try {
             board = boardService.findById(id);
         } catch (Exception e) {
@@ -84,7 +102,10 @@ public class BoardController {
 
     @ApiOperation(value = "게시글 수정")
     @PutMapping("/board")
-    public ResponseEntity<?> updateBoard(@RequestBody Board board) {
+    public ResponseEntity<?> updateBoard(@RequestBody Board board, Authentication authentication) {
+        if (isInValidAuthentication(authentication)) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         try {
             boardService.update(board);
         } catch (Exception e) {
@@ -96,13 +117,38 @@ public class BoardController {
 
     @ApiOperation(value = "게시글 삭제")
     @DeleteMapping("/board/{id}")
-    public ResponseEntity<?> deleteBoard(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteBoard(@PathVariable("id") Long id, Authentication authentication) {
+        if (isInValidAuthentication(authentication)) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         try {
             boardService.delete(id);
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    @ApiOperation(value = "게시글 좋아요(증가,감소)")
+    @PutMapping("/board/likes")
+    public ResponseEntity<?> likesUpdateBoard(@RequestBody Map<String, Object> map, Authentication authentication) {
+
+        if (isInValidAuthentication(authentication)) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        Claims claims = (Claims) authentication.getPrincipal();
+
+        Long userId = claims.get("userId", Long.class);
+        int likes = (int) map.get("likes");
+        try {
+            boardService.updateLikes(userId, likes);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         return new ResponseEntity<>(HttpStatus.OK);
     }
